Added ansigenome_managed block to README templates.
Like with templating in Ansible where every generated file should have
an `{{ ansible_managed }}` comment at the top of the file, the
generated README should have a hint as well.
I know there is a hint at the bottom of the file but I think it makes
sense to have a comment at the top as well so when someone opens the
README file they immediately see that they are not editing the source
file.

I tested the Markdown comment with GitHub and Ansible Galaxy:

* https://github.com/ypid/ansible-initramfs_message#initramfs_message
* https://galaxy.ansible.com/list#/roles/2807