Do not sort/mix the order of role dependencies.
It is expected that the order of role dependencies has at least some
meaning so that the sorting applied by the author of the role is more
important than sorting by Ansigenome or as it was done before this
commit by applying a kind of random order (dict data structure sorting)
to it.

Also use a more Ansible`ish` Jinja2 by using common filters from Jinja2
`map` and a custom one from Ansible `unique`.